requirements.txt
cog>=0.15.9
torch
transformers
stable-audio-tools
python-dotenv
huggingface_hub 